+++ src/main/java/com/rongzm/mapper/OperateLogMapper.java
@@ -3,13 +3,13 @@ package com.rongzm.mapper;
 import com.rongzm.entity.OperateLog;
 
 public interface OperateLogMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(OperateLog record);
 
     int insertSelective(OperateLog record);
 
-    OperateLog selectByPrimaryKey(Long id);
+    OperateLog selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(OperateLog record);
